refactor(server): rename files and update exporting standard
