+++ server/src/app/http/middlewares/cors.ts
@@ -1,8 +1,6 @@
 import cors from 'cors';
 import type express from 'express';
 
-function register(app: express.Application): void {
+export function register(app: express.Application): void {
   app.use(cors());
 }
-
-export default register;
