+++ server/src/app/http/middlewares/index.ts
@@ -1,0 +1,3 @@
+export { register as registerCORS } from './cors';
+export { register as registerJsonParser } from './json-parser';
+export { register as registerLogger } from './logger';
