+++ server/src/app/http/middlewares/json-parser.ts
@@ -1,0 +1,5 @@
+import express from 'express';
+
+export function register(app: express.Application): void {
+  app.use(express.json());
+}
