+++ server/src/app/http/middlewares/logger.ts
@@ -1,8 +1,6 @@
 import type express from 'express';
 import morgan from 'morgan';
 
-function register(app: express.Application): void {
+export function register(app: express.Application): void {
   app.use(morgan('dev'));
 }
-
-export default register;
