+++ server/src/app/modules/classSchedules/index.ts
@@ -1,0 +1,1 @@
+export { register as registerClassSchedulesRoutes } from './routes';
