+++ server/src/app/modules/classes/index.ts
@@ -1,0 +1,1 @@
+export { register as registerClassesRoutes } from './routes';
