+++ server/src/app/modules/connections/count/count-connections-controller.ts
@@ -3,10 +3,8 @@ import HttpStatus from 'http-status-codes';
 import database from '../../../../database/connection';
 import { type Controller } from '../../../http';
 
-const getConnectionsCountController: Controller = async (_request, response) => {
+export const countConnectionsController: Controller = async (_request, response) => {
   const [{ total }] = await database('connections').count('* as total');
 
   response.status(HttpStatus.OK).json({ total });
 };
-
-export default getConnectionsCountController;
