+++ server/src/app/modules/connections/count/index.ts
@@ -1,0 +1,1 @@
+export * from './count-connections-controller';
