+++ server/src/app/modules/connections/create/create-connection-controller.ts
@@ -3,12 +3,10 @@ import HttpStatus from 'http-status-codes';
 import database from '../../../../database/connection';
 import { type Controller } from '../../../http';
 
-const createConnectionController: Controller = async (request, response) => {
+export const createConnectionController: Controller = async (request, response) => {
   const { user: user_id } = request.body;
 
   await database('connections').insert({ user_id });
 
   response.status(HttpStatus.CREATED).send();
 };
-
-export default createConnectionController;
