--- server/src/app/modules/connections/getCount/index.ts
@@ -1,1 +1,0 @@
-export { default as getConnectionsCountController } from './getConnectionsCount.controller';
