+++ server/src/app/modules/connections/index.ts
@@ -1,0 +1,1 @@
+export { register as registerConnectionsRoutes } from './routes';
