+++ server/src/app/modules/users/index.ts
@@ -1,0 +1,1 @@
+export { register as registerUsersRoutes } from './routes';
