+++ server/src/app/modules/users/routes.ts
@@ -1,9 +1,7 @@
 import express from 'express';
 
-function register(prefix: string, appRouter: express.Router): void {
+export function register(prefix: string, appRouter: express.Router): void {
   const router = express.Router();
 
   appRouter.use(prefix, router);
 }
-
-export default register;
