--- server/src/middlewares/index.ts
@@ -1,3 +1,0 @@
-export { default as registerCORS } from './cors';
-export { default as registerJsonParser } from './jsonParser';
-export { default as registerLogger } from './logger';
