--- server/src/middlewares/jsonParser.ts
@@ -1,7 +1,0 @@
-import express from 'express';
-
-function register(app: express.Application): void {
-  app.use(express.json());
-}
-
-export default register;
